Fix build error in Android
external/f2fs-tools/mkfs/f2fs_format.c:392:31: error: format specifies type 'unsigned long long' but the argument has type 'size_t' (aka 'unsigned int') [-Werror,-Wformat]
  391 |                         MSG(1, "\tError: Unaligned segment0 start (%u) for zoned LU (zone_blocks: %"PRIu64")\n",
      |                         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  392 |                                 get_sb(segment0_blkaddr), c.zone_blocks);
      |                                 ~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~
external/f2fs-tools/include/f2fs_fs.h:289:18: note: expanded from macro 'MSG'
  289 |                         printf(fmt, ##__VA_ARGS__);                     \
      |                                ~~~    ^~~~~~~~~~~
external/f2fs-tools/mkfs/f2fs_format.c:404:31: error: format specifies type 'unsigned long long' but the argument has type 'size_t' (aka 'unsigned int') [-Werror,-Wformat]
  402 |                         MSG(1, "\tError: Unaligned start (%"PRIu64") for zoned LU from segment0 (%u) (zone_blocks: %"PRIu64")\n",
      |                         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  403 |                                 c.devices[1].start_blkaddr,
      |                                 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
  404 |                                 get_sb(segment0_blkaddr), c.zone_blocks);
      |                                 ~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~
external/f2fs-tools/include/f2fs_fs.h:289:18: note: expanded from macro 'MSG'
  289 |                         printf(fmt, ##__VA_ARGS__);                     \
      |                                ~~~    ^~~~~~~~~~~

+++ mkfs/f2fs_format.c
@@ -389,7 +389,8 @@ static int f2fs_prepare_super_block(void)
 			 * aligned at the zone.
 			 */
 			MSG(1, "\tError: Unaligned segment0 start (%u) for zoned LU (zone_blocks: %"PRIu64")\n",
-				get_sb(segment0_blkaddr), c.zone_blocks);
+				get_sb(segment0_blkaddr),
+				(uint64_t)c.zone_blocks);
 			return -1;
 		} else if (c.ndevs > 1 &&
 			(c.devices[1].start_blkaddr - get_sb(segment0_blkaddr)) % c.zone_blocks) {
@@ -401,7 +402,8 @@ static int f2fs_prepare_super_block(void)
 			 */
 			MSG(1, "\tError: Unaligned start (%"PRIu64") for zoned LU from segment0 (%u) (zone_blocks: %"PRIu64")\n",
 				c.devices[1].start_blkaddr,
-				get_sb(segment0_blkaddr), c.zone_blocks);
+				get_sb(segment0_blkaddr),
+				(uint64_t)c.zone_blocks);
 			return -1;
 		}
 	}
